Restructure commands and get database working
Structure command files in a more scalable fashion. Create a placeholder
name command with a dummy database query to ensure our database
connection works.

Later, the name command will allow a user to register with their real
name and the courses in which they're enrolled, and if they're a
professor or not.